[feat] 중복 닉네임 테스트 코드
