+++ src/main/java/com/lof/lofserver/service/user/validate/UserValidateImpl.java
@@ -26,5 +26,4 @@ public class UserValidateImpl implements UserValidate {
             throw new UserException(UserExceptionType.NICKNAME_ALREADY_EXIST);
         });
     }
-
 }
