+++ src/test/java/com/lof/lofserver/service/user/UserServiceImplTest.java
@@ -6,6 +6,7 @@ import com.lof.lofserver.exception.UserException;
 import com.lof.lofserver.exception.UserExceptionType;
 import com.lof.lofserver.service.user.request.UserSavedInfo;
 import com.lof.lofserver.service.user.response.UserResponseInfo;
+import com.lof.lofserver.service.user.validate.UserValidateImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,7 +17,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
@@ -28,6 +28,9 @@ class UserServiceImplTest {
 
     @Mock
     private UserRepository userRepository;
+
+    @Mock
+    private UserValidateImpl userValidateImpl;
 
     @Test
     @DisplayName("유저 닉네임 설정하기 - 성공")
@@ -109,21 +112,5 @@ class UserServiceImplTest {
         assertThat(userResponseInfo).isNotNull();
         assertThat(userResponseInfo.isNewUser()).isFalse();
     }
-
-    @Test
-    @DisplayName("유저 닉네임 설정하기 - 실패(중복)")
-    void validateDuplicateUserNickname() {
-        //given
-        String user1Nickname = "user1";
-        String user2Nickname = "user1";
-
-        //when
-        given(userRepository.findByNickname(user1Nickname)).willReturn(Optional.of(UserEntity.builder().nickname(user1Nickname).build()));
-
-        //then
-        UserException throwable = assertThrows(UserException.class, () -> userService.setUserNickName(1L, user2Nickname));
-        assertThat(throwable.getExceptionType()).isEqualTo(UserExceptionType.NICKNAME_ALREADY_EXIST);
-    }
-
 
 }
